refactor(server): retrieval trace の DB 操作を repository へ寄せる
